Lighting System
A designer friendly lighting system which can be adjusted through a dropdown box accessed through the details panel. the spotlight has a total of 4 modes: static, pulsing, following the player (while remaining in its location) and moving with the player. within the plug in there is also a trigger box that allows the light to switch modes in play whenever the player walks through it.